# platform/wanwu/docs/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "## HTTP Header\r\n| Header        | 说明      |\r\n| ------------- | --------- |\r\n| Authorization | JWT token |\r\n| X-Language    | 语言Code  |\r\n| X-Org-Id      | 组织ID    |\r\n\r\n## HTTP Status\r\n| HTTP Status             | 说明                   |\r\n| ----------------------- | ---------------------- |\r\n| 200, StatusOK           | 请求返回成功           |\r\n| 400, StatusBadRequest   | 请求返回失败，用于业务 |\r\n| 401, StatusUnauthorized | JWT认证失败            |\r\n| 403, StatusForbidden    | 没有权限               |\r\n\r\n## 权限-菜单对应表\r\n| 一级权限        | 二级权限  | 三级权限 | 一级菜单 | 二级菜单 | 三级菜单 |\r\n|-------------|-------|------|------|------|------|\r\n| guest       |       |      | 【访客】 |      |      |\r\n| common      |       |      | 【通用】 |      |      |\r\n| permission  |       |      | 权限管理 |      |      |\r\n| permission  | user  |      | 权限管理 | 用户管理 |      |\r\n| permission  | org   |      | 权限管理 | 组织管理 |      |\r\n| permission  | role  |      | 权限管理 | 角色管理 |      |\r\n\r\n## `/v1/user/permission`返回用例\r\n```json\r\n{\r\n  \"code\": 0,\r\n  \"data\": {\r\n    \"orgPermission\": {\r\n      \"org\": {\"id\": \"test-org-id\", \"name\": \"test-org-name\"},\r\n      \"permissions\": [\r\n        {\"perm\": \"permission\"},\r\n        {\"perm\": \"permission.user\"},\r\n        {\"perm\": \"permission.org\"},\r\n        {\"perm\": \"permission.role\"}\r\n      ]\r\n    }\r\n  },\r\n  \"msg\": \"操作成功\"\r\n}\r\n```",
...
        "contact": {},
...
    "basePath": "/v1",
    "paths": {
...
        "/model": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "model"
                ],
                "summary": "‌查询单个模型",
                "parameters": [
                    {
                        "enum": [
                            "llm",
                            "embedding",
                            "rerank"
                        ],
                        "type": "string",
                        "description": "模型类型",
                        "name": "modelType",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "模型名称",
                        "name": "model",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "模型供应商",
                        "name": "provider",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.ModelInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "model"
                ],
                "summary": "导入模型更新",
                "parameters": [
                    {
                        "description": "模型变更信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ImportOrUpdateModelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "第三方模型的导入",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "model"
                ],
                "summary": "模型导入",
                "parameters": [
                    {
                        "description": "导入模型信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ImportOrUpdateModelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "model"
                ],
                "summary": "导入模型删除",
                "parameters": [
                    {
                        "description": "模型删除key",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.DeleteModelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/model/list": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "model"
                ],
                "summary": "导入模型列表",
                "parameters": [
                    {
                        "enum": [
                            "llm",
                            "embedding",
                            "rerank"
                        ],
                        "type": "string",
                        "description": "模型类型",
                        "name": "modelType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "模型供应商",
                        "name": "provider",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "模型显示名称",
                        "name": "displayName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "启用状态（true: 启用）",
                        "name": "isActive",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/response.ListResult"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "$ref": "#/definitions/response.ModelBrief"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/model/select/embedding": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "model"
                ],
                "summary": "embedding模型列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/response.ListResult"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "$ref": "#/definitions/response.ModelBrief"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/model/select/llm": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "model"
                ],
                "summary": "llm模型列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/response.ListResult"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "$ref": "#/definitions/response.ModelBrief"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/model/select/rerank": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "model"
                ],
                "summary": "rerank模型列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/response.ListResult"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "$ref": "#/definitions/response.ModelBrief"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/model/status": {
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "model"
                ],
                "summary": "模型启用/关闭",
                "parameters": [
                    {
                        "description": "启用/关闭 的模型信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ModelStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "mp.ProviderModelByOpenAICompatible": {
            "type": "object",
            "properties": {
                "embedding": {
                    "$ref": "#/definitions/mp_openai_compatible.Embedding"
                },
                "llm": {
                    "$ref": "#/definitions/mp_openai_compatible.LLM"
                },
                "rerank": {
                    "$ref": "#/definitions/mp_openai_compatible.Rerank"
                }
            }
        },
        "mp.ProviderModelByYuanjing": {
            "type": "object",
            "properties": {
                "embedding": {
                    "$ref": "#/definitions/mp_yuanjing.Embedding"
                },
                "llm": {
                    "$ref": "#/definitions/mp_yuanjing.LLM"
                },
                "rerank": {
                    "$ref": "#/definitions/mp_yuanjing.Rerank"
                }
            }
        },
        "mp.ProviderModelConfig": {
            "type": "object",
            "properties": {
                "providerOpenAICompatible": {
                    "$ref": "#/definitions/mp.ProviderModelByOpenAICompatible"
                },
                "providerYuanJing": {
                    "$ref": "#/definitions/mp.ProviderModelByYuanjing"
                }
            }
        },
        "mp_openai_compatible.Embedding": {
            "type": "object",
            "properties": {
                "apiKey": {
                    "description": "ApiKey",
                    "type": "string"
                },
                "endpointUrl": {
                    "description": "推理url",
                    "type": "string"
                }
            }
        },
        "mp_openai_compatible.LLM": {
            "type": "object",
            "properties": {
                "apiKey": {
                    "description": "ApiKey",
                    "type": "string"
                },
                "endpointUrl": {
                    "description": "推理url",
                    "type": "string"
                },
                "functionCalling": {
                    "description": "函数调用是否支持",
                    "type": "string",
                    "enum": [
                        "noSupport",
                        "toolCall",
                        "functionCall"
                    ]
                }
            }
        },
...
        "mp_openai_compatible.Rerank": {
            "type": "object",
            "properties": {
                "apiKey": {
                    "description": "ApiKey",
                    "type": "string"
                },
                "endpointUrl": {
                    "description": "推理url",
                    "type": "string"
                }
            }
        },
        "mp_yuanjing.Embedding": {
            "type": "object",
            "properties": {
                "apiKey": {
                    "description": "ApiKey",
                    "type": "string"
                },
                "endpointUrl": {
                    "description": "推理url",
                    "type": "string"
                }
            }
        },
        "mp_yuanjing.LLM": {
            "type": "object",
            "properties": {
                "apiKey": {
                    "description": "ApiKey",
                    "type": "string"
                },
                "endpointUrl": {
                    "description": "推理url",
                    "type": "string"
                },
                "functionCalling": {
                    "description": "函数调用是否支持",
                    "type": "string",
                    "enum": [
                        "noSupport",
                        "toolCall",
                        "functionCall"
                    ]
                }
            }
        },
...
        "request.Avatar": {
            "type": "object",
            "properties": {
                "key": {
                    "description": "前端透传给后端用于保存avatar，例如：custom-upload/avatar/abc/def.png",
                    "type": "string"
                },
                "path": {
                    "description": "前端请求地址，例如：/v1/static/avatar/abc/def.png (请求非必填)",
                    "type": "string"
                }
            }
        },
...
        "request.DeleteModelRequest": {
            "type": "object",
            "required": [
                "model",
                "modelType",
                "provider"
            ],
            "properties": {
                "model": {
                    "description": "模型名称",
                    "type": "string"
                },
                "modelType": {
                    "description": "模型类型",
                    "type": "string",
                    "enum": [
                        "llm",
                        "embedding",
                        "rerank"
                    ]
                },
                "provider": {
                    "description": "模型供应商",
                    "type": "string"
                }
            }
        },
...
        "request.ImportOrUpdateModelRequest": {
            "type": "object",
            "required": [
                "model",
                "modelType",
                "provider"
            ],
            "properties": {
                "avatar": {
                    "description": "模型图标路径",
                    "allOf": [
                        {
                            "$ref": "#/definitions/request.Avatar"
                        }
                    ]
                },
                "config": {},
                "displayName": {
                    "description": "模型显示名称",
                    "type": "string"
                },
                "examples": {
                    "description": "仅用于swagger展示；模型对应供应商中的对应llm、embedding或rerank结构是config实际的参数",
                    "allOf": [
                        {
                            "$ref": "#/definitions/mp.ProviderModelConfig"
                        }
                    ]
                },
                "model": {
                    "description": "模型名称",
                    "type": "string"
                },
                "modelId": {
                    "type": "string"
                },
                "modelType": {
                    "description": "模型类型",
                    "type": "string",
                    "enum": [
                        "llm",
                        "embedding",
                        "rerank"
                    ]
                },
                "provider": {
                    "description": "模型供应商",
                    "type": "string",
                    "enum": [
                        "OpenAI-API-compatible",
                        "YuanJing"
                    ]
                },
                "publishDate": {
                    "description": "模型发布时间",
                    "type": "string"
                }
            }
        },
...
        "request.ModelStatusRequest": {
            "type": "object",
            "required": [
                "model",
                "modelType",
                "provider"
            ],
            "properties": {
                "isActive": {
                    "description": "启用状态（true: 启用，false: 禁用）",
                    "type": "boolean"
                },
                "model": {
                    "description": "模型名称",
                    "type": "string"
                },
                "modelType": {
                    "description": "模型类型",
                    "type": "string",
                    "enum": [
                        "llm",
                        "embedding",
                        "rerank"
                    ]
                },
                "provider": {
                    "description": "模型供应商",
                    "type": "string"
                }
            }
        },
...
        "response.ListResult": {
            "type": "object",
            "properties": {
                "list": {},
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "response.ModelBrief": {
            "type": "object",
            "required": [
                "model",
                "modelType",
                "provider"
            ],
            "properties": {
                "avatar": {
                    "description": "模型图标路径",
                    "allOf": [
                        {
                            "$ref": "#/definitions/request.Avatar"
                        }
                    ]
                },
                "createdAt": {
                    "type": "string"
                },
                "displayName": {
                    "description": "模型显示名称",
                    "type": "string"
                },
                "isActive": {
                    "description": "启用状态（true: 启用，false: 禁用）",
                    "type": "boolean"
                },
                "model": {
                    "description": "模型名称",
                    "type": "string"
                },
                "modelId": {
                    "type": "string"
                },
                "modelType": {
                    "type": "string",
                    "enum": [
                        "llm",
                        "embedding",
                        "rerank"
                    ]
                },
                "orgId": {
                    "type": "string"
                },
                "provider": {
                    "description": "模型供应商",
                    "type": "string",
                    "enum": [
                        "OpenAI-API-compatible",
                        "YuanJing"
                    ]
                },
                "publishDate": {
                    "description": "模型发布时间",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "response.ModelInfo": {
            "type": "object",
            "required": [
                "model",
                "modelType",
                "provider"
            ],
            "properties": {
                "avatar": {
                    "description": "模型图标路径",
                    "allOf": [
                        {
                            "$ref": "#/definitions/request.Avatar"
                        }
                    ]
                },
                "config": {},
                "createdAt": {
                    "type": "string"
                },
                "displayName": {
                    "description": "模型显示名称",
                    "type": "string"
                },
                "examples": {
                    "description": "仅用于swagger展示；模型对应供应商中的对应llm、embedding或rerank结构是config实际的参数",
                    "allOf": [
                        {
                            "$ref": "#/definitions/mp.ProviderModelConfig"
                        }
                    ]
                },
                "isActive": {
                    "description": "启用状态（true: 启用，false: 禁用）",
                    "type": "boolean"
                },
                "model": {
                    "description": "模型名称",
                    "type": "string"
                },
                "modelId": {
                    "type": "string"
                },
                "modelType": {
                    "type": "string",
                    "enum": [
                        "llm",
                        "embedding",
                        "rerank"
                    ]
                },
                "orgId": {
                    "type": "string"
                },
                "provider": {
                    "description": "模型供应商",
                    "type": "string",
                    "enum": [
                        "OpenAI-API-compatible",
                        "YuanJing"
                    ]
                },
                "publishDate": {
                    "description": "模型发布时间",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}